+++ tests/test-distribution/src/test/java/org/eclipse/jetty/tests/distribution/DynamicListenerTests.java
@@ -21,6 +21,7 @@ import java.util.concurrent.TimeUnit;
 
 import org.eclipse.jetty.client.api.ContentResponse;
 import org.eclipse.jetty.http.HttpStatus;
+import org.eclipse.jetty.tests.hometester.JettyHomeTester;
 import org.eclipse.jetty.util.IO;
 import org.junit.jupiter.api.Test;
